% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffOriginal.R
\name{diffOriginal}
\alias{diffOriginal}
\title{Visual diff of original version of a file to the latest version.}
\usage{
diffOriginal(file, side_by_side = TRUE, ignore_white_space = FALSE)
}
\arguments{
\item{file}{file path from working directory}

\item{side_by_side}{Logical. Should diffs be displayed side by side?}

\item{ignore_white_space}{Logical. Should white space be ignored?}
}
\description{
Compares the latest version of a file to the initial version checked into
the repository. The output will appear in the viewer and only rows where
there have been additions, deletions or modifications in the file will be shown.
}
\examples{
with_demoRepo({
  diffOriginal("script/data-assembly.R")
})

}
