% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffPreviousVersions.R
\name{diffPreviousVersions}
\alias{diffPreviousVersions}
\title{Visual diff of previous version of a file to the latest version.}
\usage{
diffPreviousVersions(
  file,
  previous_version,
  current_version = NULL,
  banner_new = NULL,
  banner_prev = NULL,
  side_by_side = TRUE,
  ignore_white_space = FALSE
)
}
\arguments{
\item{file}{file path from working directory}

\item{previous_version}{commit hash of version to compare to current revision}

\item{current_version}{current version (defaults to local copy)}

\item{banner_new}{Header for first file in viewer}

\item{banner_prev}{Header for second file in viewer}

\item{side_by_side}{Logical. Should diffs be displayed side by side?}

\item{ignore_white_space}{Logical. Should white space be ignored?}
}
\description{
Compares the current version of a script with a previous revision.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the script will be shown.

The user can also provide two previous versions
}
\examples{
with_demoRepo({

 prev_version <- gitLog("script/data-assembly.R")[2,][["last_commit"]]
 
 diffPreviousVersions(file = "script/data-assembly.R", 
                       previous_version = prev_version)
})

}
