% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffQced.R
\name{diffQced}
\alias{diffQced}
\title{Visual diff of last QCed version of a file to the local version.}
\usage{
diffQced(file)
}
\arguments{
\item{file}{file path from working directory}
}
\description{
Compares the local version of a file with the most recent QCed version.
The output will appear in the viewer and only rows where there have been
additions, deletions or modifications in the file will be shown.
}
\examples{
with_demoRepo({
  diffQced("script/data-assembly.R")
})

}
