% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitLog.R
\name{gitLog}
\alias{gitLog}
\title{Get commit info of last modified version of a file}
\usage{
gitLog(file)
}
\arguments{
\item{file}{character file path (either the absolute or relative file path from the QC log)}
}
\description{
Returns a list including the commit hash, last author and datetime of last
commit for a specific file.
}
\examples{
with_demoRepo({
  gitLog("script/examp-txt.txt")
})

}
