% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAccept.R
\name{logAccept}
\alias{logAccept}
\title{Approve a File}
\usage{
logAccept(file)
}
\arguments{
\item{file}{file to accept in the QC log}
}
\description{
Add a row to the QC log accepting a file, recording the commit hash of the
commit where the file's last modification occurred.
}
