% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAssign.R
\name{logAssign}
\alias{logAssign}
\title{Assign Files to QC Log}
\usage{
logAssign(file, reviewer = "anyone")
}
\arguments{
\item{file}{character file path (either the absolute or relative file path from the QC log)}

\item{reviewer}{specify a specific person to review the file (defaults to "anyone")}
}
\description{
Adds a row to the QC log for the file given with the default reviewer
set to "anyone".
}
\details{
The purpose of \code{logAssign()} is to indicate a file is ready for QC. This
will allow the file to appear on the output of \code{logPending()}.
}
\examples{
\dontrun{ 
setwd(demoRepo)
logAssign(file = "script/examp-yaml.yaml", reviewer = "person1")
}

}
