% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logCreate.R
\name{logCreate}
\alias{logCreate}
\title{Create a QC Log}
\usage{
logCreate()
}
\description{
Creates an empty table in the directory specified with the name "QClog.csv".
By default, the csv will be created in the directory where the R studio
project exists.
}
