% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logPending.R, R/logSummary.R
\name{logPending}
\alias{logPending}
\title{List Files that Require QC}
\usage{
logPending()

logPending()
}
\description{
Checks the files in the QC log and returns those that are not up to date with
QC, meaning files that have only been assigned or have had modifications since
their last QC was completed.

Checks the files in the QC log and returns their QC status. Statuses include:
"Assigned, needs QC", "Modified, needs QC", and "Fully QCed".
}
