% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logSummary.R
\name{logSummary}
\alias{logSummary}
\title{Returns QC Status of Files in QC log}
\usage{
logSummary()
}
\description{
Checks the files in the QC log and returns their QC status. Statuses include:
"Assigned, needs QC", "Modified, needs QC", and "Fully QCed".
}
\examples{
\dontrun{ 
setwd(demoRepo)
logSummary()
}

}
