% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qctoolsPackage.R
\docType{package}
\name{qctools-package}
\alias{qctools}
\alias{qctools-package}
\title{qctools: QC Management and Helpers}
\description{
Create, append and review the QC log. Provides helper functions for completing QC tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/metrumresearchgroup/qctools}
  \item \url{https://metrumresearchgroup.github.io/qctools}
  \item Report bugs at \url{https://github.com/metrumresearchgroup/qctools/issues}
}

}
\author{
\strong{Maintainer}: Michael McDermott \email{michaelm@metrumrg.com}

Authors:
\itemize{
  \item Eric Anderson \email{andersone@metrumrg.com}
}

}
\keyword{internal}
