% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderQCSummary.R
\name{renderQCSummary}
\alias{renderQCSummary}
\title{Render a QC Summary Document}
\usage{
renderQCSummary(.output_dir = NULL)
}
\arguments{
\item{.output_dir}{Character string (optional). Path to the directory where the output PDF
should be saved. If not provided, the document will not be saved locally.}
}
\value{
Invisible. The function will save a PDF report named "qc-summary-(current date).pdf"
to the specified output directory (if available). If the R session is interactive, it will also
open the PDF in the default browser.
}
\description{
This function generates a QC summary based on scripts within a given project
and renders it into a PDF report.
}
\details{
A high level summary of all scripts that need to be QCed will be displayed along
with a summary of recently modified scripts that have not been assigned to the QC log.
}
\examples{
\dontrun{
  with_demoRepo({
    renderQCSummary()
  })
}

}
