% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repoHistory.R
\name{repoHistory}
\alias{repoHistory}
\title{Get Git Repository History}
\usage{
repoHistory()
}
\description{
Returns a data frame containing every commit made to the Git repository.
The information recorded includes the commit hash, author, date, and
files changed in each commit.
}
\keyword{internal}
