% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcsLastCommit.R
\name{vcsLastCommit}
\alias{vcsLastCommit}
\title{Get commit info of last modified version of a file}
\usage{
vcsLastCommit(list_of_files)
}
\arguments{
\item{list_of_files}{vector of file paths (can also pass only 1 file)}
}
\description{
Returns a data.frame including the commit hash, last author and datetime of last
commit for all input files.
}
\examples{
with_demoRepo({
  vcsLastCommit("script/examp-txt.txt")
})

}
