% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAccept.R
\name{logAccept}
\alias{logAccept}
\title{Approve a File in the QC log}
\usage{
logAccept(file)
}
\arguments{
\item{file}{character file path (either the absolute or relative file path from the QC log)}
}
\description{
Add a row to the QC log accepting a file, recording the commit hash of the
commit where the file's last modification occurred. The username of the
reviewer will be included in the new row as well.
}
\examples{
with_demoRepo({
  logAccept(file = "script/data-assembly.R")
})

}
