% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAssign.R
\name{logAssign}
\alias{logAssign}
\title{Assign Files to QC Log}
\usage{
logAssign(file = tempfile())
}
\arguments{
\item{file}{character file path (either the absolute or relative file path from the QC log)}

\item{reviewer}{specify a specific person to review the file}
}
\description{
Appends a row to the QC log for the file given with the default reviewer
set to "anyone".
}
