% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logCreate.R
\name{logCreate}
\alias{logCreate}
\title{Create a QC Log}
\usage{
logCreate()
}
\description{
Creates an empty QC log file with the name "QClog.csv". This file will be
created in the directory where the RStudio project file (\verb{*.Rproj}) exists.
}
\examples{
\dontrun{ 
logCreate()
}

}
